MB-69666: Fix -Werror,-Wshorten-64-to-32 reported on Ubuntu
When building with Clang 18 on Ubuntu 24 there is various
errors reported caused by passing size_t to htonl()

Reviewed-on: https://review.couchbase.org/c/couchstore/+/237266
Tested-by: Build Bot <build@couchbase.com>

+++ src/stream.cc
@@ -449,7 +449,7 @@ void EncryptedStream::write_chunk() {
         return;
     }
     auto encrypted = cipher->encrypt(chunk);
-    uint32_t cipher_len = htonl(encrypted.size());
+    uint32_t cipher_len = htonl(gsl::narrow_cast<uint32_t>(encrypted.size()));
     underlying->write(
             {reinterpret_cast<char*>(&cipher_len), sizeof(cipher_len)});
     underlying->write(encrypted);
